First Commit
This database was my first big self-driven project for college. In the first semester we were to build a database on a thing we were interested in building. I play lots of General Manager Simulators for the NHL, so I thought a database like one I use in game would be appropriate. It was a fun experience, with a lot of lessons along the way.